Crate c509

C509 certificate CLI

Structs

C509Json 🔒

A struct representing the JSON format of C509 certificate.

Enums

Cli 🔒

Commands for C509 certificate generation, verification and decoding

Constants

ED25519 🔒

Ed25519 oid and parameter - default algorithm.

SELF_SIGNED_INT 🔒

Integer indicate that certificate is self-signed. 0 for Natively Signed C509 Certificate following X.509 v3 1 for CBOR re-encoding of X.509 v3 Certificate

Functions

decode 🔒

Decode the certificate to JSON.

determine_issuer 🔒

Determine issuer of the certificate. If self-signed is true, issuer is the same as subject. Otherwise, issuer must be present.

extract_relative_distinguished_name 🔒

Extract a RelativeDistinguishedName from a Name.

generate 🔒

A function to generate C509 certificate.

get_key_type 🔒

Get the key type. Currently support only Ed25519.

main 🔒

parse_or_default_date 🔒

Parse date string to i64.

parse_public_key 🔒

Parse public key from file path.

parse_serial_number 🔒

Generate random serial number if not provided

time_to_string 🔒

Convert time in i64 to string.

validate_certificate_type 🔒

Validate the certificate type.

verify 🔒

Verify the signature of the certificate given public key file path.

write_to_output_file 🔒

Write a data to a file given an output path.